Fixed bug with Double to Int

+++ src/main/java/commands/ChessCommand.java
@@ -229,10 +229,10 @@ public class ChessCommand {
                     state.getPrevElo().put(blackSidePlayer.discordId, blackSidePlayer.elo);
                     double whiteSideWinChance = EloRanking.calculateProbabilityOfWin(whiteSidePlayer.elo, blackSidePlayer.elo);
                     reply = String.format("`Starting Chess Match!` %s (elo: %d, odds: %s%%) vs. %s (elo: %d, odds: %s%%)\n%s will go first...", whiteSidePlayer.name,
-                                                                                                                      whiteSidePlayer.elo,
+                                                                                                                    (int)whiteSidePlayer.elo,
                                                                                                                       formatPercent.format(whiteSideWinChance*100),
                                                                                                                       blackSidePlayer.name,
-                                                                                                                      blackSidePlayer.elo,
+                                                                                                                    (int)blackSidePlayer.elo,
                                                                                                                       formatPercent.format((1-whiteSideWinChance)*100),
                                                                                                                       whiteSidePlayer.name);
                     gameType = GameType.CVP;
@@ -276,10 +276,10 @@ public class ChessCommand {
                     if (message.equalsIgnoreCase("y")) {
                         double whiteSideWinChance = EloRanking.calculateProbabilityOfWin(whiteSidePlayer.elo, blackSidePlayer.elo);
                         reply = String.format("`Starting Chess Match!` %s (elo: %d, odds: %s%%) vs. %s (elo: %d, odds: %s%%)", whiteSidePlayer.name,
-                                                                                                                                whiteSidePlayer.elo,
+                                                                                                                                (int)whiteSidePlayer.elo,
                                                                                                                                 formatPercent.format(whiteSideWinChance*100),
                                                                                                                                 blackSidePlayer.name,
-                                                                                                                                blackSidePlayer.elo,
+                                                                                                                                (int)blackSidePlayer.elo,
                                                                                                                                 formatPercent.format((1-whiteSideWinChance)*100),
                                                                                                                                 whiteSidePlayer.name);
                         boardImageFile = new File(GAME_BOARD_IMAGE_LOCATION);
